+++ Editor/Sources/Recorders/_Inputs/Camera/CameraInputSettings.cs
@@ -1,6 +1,7 @@
 using System;
 using System.Collections.Generic;
 using System.ComponentModel;
+using System.Linq;
 using UnityEngine;
 using UnityEngine.Rendering;
 
@@ -73,8 +74,26 @@ namespace UnityEditor.Recorder.Input
         protected internal override void CheckForErrors(List<string> errors)
         {
             base.CheckForErrors(errors);
-            if (Source == ImageSource.TaggedCamera && string.IsNullOrEmpty(CameraTag))
-                errors.Add("Missing tag for camera selection");
+            if (Source == ImageSource.TaggedCamera)
+            {
+                if (string.IsNullOrEmpty(CameraTag))
+                    errors.Add("Missing tag for camera selection");
+                else
+                {
+                    try
+                    {
+                        var objs = GameObject.FindGameObjectsWithTag(CameraTag);
+                        var cams = objs.Select(obj => obj.GetComponent<Camera>()).Where(c => c != null);
+
+                        if (cams.Count() == 0)
+                            errors.Add("No camera has the requested target tag '" + CameraTag + "'");
+                    }
+                    catch (UnityException)
+                    {
+                        errors.Add("The requested target tag '" + CameraTag + "' does not exist in the project");
+                    }
+                }
+            }
         }
     }
 }